Consolidate serializers and views in api app
